+++ contrib/runners/python_runner/tests/unit/test_pythonrunner.py
@@ -22,7 +22,6 @@ from oslo_config import cfg
 
 from python_runner import python_runner
 from st2actions.container.base import RunnerContainer
-from st2common.runners.python_action_wrapper import PythonActionWrapper
 from st2common.runners.base_action import Action
 from st2common.runners.utils import get_action_class_instance
 from st2common.services import config as config_service
@@ -32,6 +31,7 @@ from st2common.constants.action import LIVEACTION_STATUS_TIMED_OUT
 from st2common.constants.action import MAX_PARAM_LENGTH
 from st2common.constants.pack import SYSTEM_PACK_NAME
 from st2common.persistence.execution import ActionExecutionOutput
+from python_runner.python_action_wrapper import PythonActionWrapper
 from st2tests.base import RunnerTestCase
 from st2tests.base import CleanDbTestCase
 from st2tests.base import blocking_eventlet_spawn
